requirements.txt
pyactiveresource==2.2.1
PyYAML==5.3.1
ShopifyAPI==7.0.3
six==1.14.0
